ZEPPELIN-3362. Unify ZeppelinContext of PythonInterpreter & IPythonInterpreter
