+++ python/src/test/resources/log4j.properties
@@ -29,3 +29,4 @@ log4j.appender.stdout.layout.ConversionPattern=%d{ABSOLUTE} %5p %c:%L - %m%n
 log4j.rootLogger=INFO, stdout
 log4j.logger.org.apache.zeppelin.python.IPythonInterpreter=DEBUG
 log4j.logger.org.apache.zeppelin.python.IPythonClient=DEBUG
+log4j.logger.org.apache.zeppelin.python=DEBUG
